+++ back/src/app.js
@@ -4,10 +4,8 @@ const mongoose = require("mongoose");
 const express = require("express");
 const cors = require("cors");
 const passport = require("passport");
-const passportLocal = require("passport-local").Strategy;
 const cookieParser = require("cookie-parser");
 const session = require("express-session");
-const flash = require("connect-flash");
 const bodyParser = require("body-parser");
 const { DB_URL, SESSION_SECRET_CODE } = process.env;
 const router = require("./routes/index");
